Update main.yml
ci: Fix "Install-Module: command not found" error in workflow

The PowerShell step in the workflow was failing with an "Install-Module: command not found" error because the commands were being run in a Bash shell.  This commit adds `shell: pwsh` to the step, ensuring that the commands are executed in a PowerShell shell, resolving the error.